PS-86 Sign in

+++ src/server/Persistify.Server.Commands/Users/GetUserCommand.cs
@@ -55,7 +55,7 @@ public class GetUserCommand : Command<GetUserRequest, GetUserResponse>
         return new GetUserResponse
         {
             Username = _user.Username,
-            Role = _user.Role
+            Role = (int)_user.Role
         };
     }
